split up specs
