+++ spec/epub/sip_writer_spec.rb
@@ -3,40 +3,8 @@
 require "spec_helper"
 require "tempfile"
 require "zip"
-require "epub_sip"
-require "yaml"
-require "pry"
-
-RSpec.shared_context "with epub fixtures" do
-  let(:input) { "#{fixture}/test.epub" }
-  let(:fixture) { File.dirname(__FILE__) + "/../support/fixtures/#{pt_objid}" }
-  let(:pt_objid) { "ark+=87302=t00000001" }
-end
-
-RSpec.describe EPUB::MetadataExtractor do
-  include_context "with epub fixtures"
-
-  let(:meta_yml) { YAML.safe_load(File.read("#{fixture}/meta.yml"), [Time]) }
-  let(:subject) { described_class.new(input) }
-
-  describe "#metadata" do
-    def compare_metadata_part(finder)
-      expect(finder.call(subject.metadata)).to eql(finder.call(meta_yml))
-    end
-
-    ["container", "mimetype", "rootfile", "manifest", "spine"].each do |subsec|
-      it "returns metadata matching the fixture's epub_contents #{subsec} " do
-        compare_metadata_part(->(x) { x["epub_contents"][subsec] })
-      end
-    end
-
-    ["creation_date", "creation_agent", "pagedata"].each do |section|
-      it "returns metadata matching the fixture's #{section} " do
-        compare_metadata_part(->(x) { x[section] })
-      end
-    end
-  end
-end
+require "epub/sip_writer"
+require_relative "./fixtures"
 
 RSpec.describe EPUB::SIPWriter do
   include_context "with epub fixtures"
